-bug#0000766: graph_tree_items ID field need to change to mediumint
git-svn-id: svn+ssh://repo.cacti.net/var/data/svnroot/cacti/branches/BRANCH_0_8_6/cacti@3667 860744bd-22fc-0310-8c96-e9fe5004b5ca

+++ cacti.sql
@@ -1797,7 +1797,7 @@ INSERT INTO graph_tree VALUES (1,1,'Default Tree');
 --
 
 CREATE TABLE graph_tree_items (
-  id smallint(5) unsigned NOT NULL auto_increment,
+  id mediumint(5) unsigned NOT NULL auto_increment,
   graph_tree_id smallint(5) unsigned NOT NULL default '0',
   local_graph_id mediumint(8) unsigned NOT NULL default '0',
   rra_id smallint(8) unsigned NOT NULL default '0',
